--- a/.github/workflows/coverage.yml
+++ b/.github/workflows/coverage.yml
@@ -26,10 +26,14 @@ jobs:
     container:
       image: 'ubuntu:22.04'
       # the default shm-size for ubuntu:18.04, but with the size increased from
-      # 65536k. github's default docker seccomp policy seems to disallow
+      # 65536k.
+      #
+      # github's default docker seccomp policy disallows
       # process_vm_readv and process_vm_writev; disable it altogether. See
       # https://docs.docker.com/engine/security/seccomp/
-      options: '--shm-size="1g" --security-opt seccomp=unconfined'
+      #
+      # by default /dev/shm doesn't support executable mappings.
+      options: '--shm-size="1g" --security-opt seccomp=unconfined --tmpfs /dev/shm:rw,nosuid,nodev,exec,size=1024g'
     env:
       CC: 'clang-12'
       CONTAINER: 'ubuntu:22.04'