--- a/src/lib/logger/rust_bindings/src/lib.rs
+++ b/src/lib/logger/rust_bindings/src/lib.rs
@@ -1,6 +1,0 @@
-// Inline the bindgen-generated bindings, suppressing warnings.
-#![allow(unused)]
-#![allow(non_upper_case_globals)]
-#![allow(non_camel_case_types)]
-#![allow(non_snake_case)]
-include!("bindings.rs");