--- a/src/lib/logger/rust_bindings/wrapper.h
+++ b/src/lib/logger/rust_bindings/wrapper.h
@@ -1,1 +1,0 @@
-#include "../logger.h"