--- a/src/lib/logger/src/lib.rs
+++ b/src/lib/logger/src/lib.rs
@@ -1,0 +1,5 @@
+mod bindings;
+
+// For legacy compatibility.
+// TODO: wrap or replace with an idiomatic rust api.
+pub use bindings::*;