--- a/src/lib/shadow-shim-helper-rs/ipc.h
+++ b/src/lib/shadow-shim-helper-rs/ipc.h
@@ -34,8 +34,8 @@ void ipcData_markPluginExited(struct IPCData* ipc_data);
 
 size_t ipcData_nbytes();
 
-void shimevent_sendEventToShadow(struct IPCData *data, const ShimEvent* e);
-void shimevent_sendEventToPlugin(struct IPCData *data, const ShimEvent* e);
+void shimevent_sendEventToShadow(struct IPCData* data, const ShimEvent* e);
+void shimevent_sendEventToPlugin(struct IPCData* data, const ShimEvent* e);
 void shimevent_recvEventFromShadow(struct IPCData* data, ShimEvent* e, bool spin);
 void shimevent_recvEventFromPlugin(struct IPCData* data, ShimEvent* e);