--- a/src/lib/shadow-shim-helper-rs/shadow_spinlock.c
+++ b/src/lib/shadow-shim-helper-rs/shadow_spinlock.c
@@ -1,4 +1,4 @@
-#include "lib/shim/shadow_spinlock.h"
+#include "shadow_spinlock.h"
 
 #include <assert.h>
 #include <sys/syscall.h>
@@ -6,9 +6,9 @@
 
 #include "lib/shim/shim_syscall.h"
 
-int shadow_spin_init(shadow_spinlock_t *lock) {
+int shadow_spin_init(shadow_spinlock_t* lock) {
     assert(lock);
-    *lock = (shadow_spinlock_t) {
+    *lock = (shadow_spinlock_t){
         ._locked = false,
     };
     return 0;
@@ -18,8 +18,9 @@ int shadow_spin_lock(shadow_spinlock_t* lock) {
     assert(lock);
     while (1) {
         bool prev_locked = atomic_load_explicit(&lock->_locked, memory_order_relaxed);
-        if (!prev_locked && atomic_compare_exchange_weak_explicit(&lock->_locked, &prev_locked, true,
-                                                        memory_order_acquire, memory_order_relaxed)) {
+        if (!prev_locked &&
+            atomic_compare_exchange_weak_explicit(
+                &lock->_locked, &prev_locked, true, memory_order_acquire, memory_order_relaxed)) {
             break;
         }
         // Always make the real syscall
@@ -28,7 +29,7 @@ int shadow_spin_lock(shadow_spinlock_t* lock) {
     return 0;
 }
 
-int shadow_spin_unlock(shadow_spinlock_t *lock) {
+int shadow_spin_unlock(shadow_spinlock_t* lock) {
     assert(lock);
     assert(atomic_load_explicit(&lock->_locked, memory_order_relaxed));
     atomic_store_explicit(&lock->_locked, false, memory_order_release);