--- a/src/lib/shadow-shim-helper-rs/shim_helper.h
+++ b/src/lib/shadow-shim-helper-rs/shim_helper.h
@@ -1,9 +1,9 @@
+
 /*
  * The Shadow Simulator
  * See LICENSE for licensing information
  */
 // clang-format off
-
 
 #ifndef shim_helpers_h
 #define shim_helpers_h