--- a/src/lib/shmem/src/lib.rs
+++ b/src/lib/shmem/src/lib.rs
@@ -1,0 +1,2 @@
+#[allow(unused)]
+mod bindings;