--- a/src/main/core/logger/shadow_logger.rs
+++ b/src/main/core/logger/shadow_logger.rs
@@ -3,7 +3,7 @@ use crate::host::host::HostInfo;
 use crate::utility::time::TimeParts;
 use crossbeam::queue::ArrayQueue;
 use log::{Level, Log, Metadata, Record, SetLoggerError};
-use log_bindings as c_log;
+use logger as c_log;
 use once_cell::sync::Lazy;
 use shadow_shim_helper_rs::emulated_time::EmulatedTime;
 use std::cell::RefCell;