--- a/src/main/host/host.h
+++ b/src/main/host/host.h
@@ -19,7 +19,7 @@ typedef struct _Host Host;
 typedef GQuark HostId;
 
 #include "lib/logger/log_level.h"
-#include "lib/shim/shim_shmem.h"
+#include "lib/shadow-shim-helper-rs/shim_shmem.h"
 #include "lib/tsc/tsc.h"
 #include "main/core/support/definitions.h"
 #include "main/host/cpu.h"