--- a/src/main/host/syscall/protected.h
+++ b/src/main/host/syscall/protected.h
@@ -15,7 +15,7 @@
 
 #include <stdbool.h>
 
-#include "lib/shim/shim_shmem.h"
+#include "lib/shadow-shim-helper-rs/shim_shmem.h"
 #include "main/host/descriptor/epoll.h"
 #include "main/host/host.h"
 #include "main/host/process.h"