--- a/src/main/host/syscall_handler.c
+++ b/src/main/host/syscall_handler.c
@@ -14,7 +14,7 @@
 #include <unistd.h>
 
 #include "lib/logger/logger.h"
-#include "lib/shim/shim_event.h"
+#include "lib/shadow-shim-helper-rs/shim_event.h"
 #include "main/bindings/c/bindings.h"
 #include "main/core/support/config_handlers.h"
 #include "main/core/worker.h"