… .github/workflows/sanitizers.yml
… cmake/ShadowTools.cmake
… config.h.in
… docs/coding.md
… src/CMakeLists.txt
… src/Cargo.lock
… src/Cargo.toml
… src/lib/CMakeLists.txt
… src/lib/logger/CMakeLists.txt
… src/lib/logger/Cargo.toml
… src/lib/logger/rust_bindings/CMakeLists.txt
… src/lib/logger/rust_bindings/Cargo.toml
… src/lib/logger/rust_bindings/README.md
… src/lib/logger/src/bindings.rs
… src/lib/shadow-build-common/Cargo.toml
… src/lib/shadow-shim-helper-rs/CMakeLists.txt
… src/lib/shadow-shim-helper-rs/shadow_sem.h
… src/lib/shadow-shim-helper-rs/shim_event.h
… src/lib/shim/CMakeLists.txt
… src/lib/shim/shim.c
… src/lib/shmem/CMakeLists.txt
… src/lib/shmem/Cargo.toml
… src/lib/shmem/src/bindings.rs
… src/lib/tsc/CMakeLists.txt
… src/lib/tsc/Cargo.toml
… src/lib/tsc/src/lib.rs
… src/main/CMakeLists.txt
… src/main/Cargo.toml
… src/main/bindings/CMakeLists.txt
… src/main/bindings/c/CMakeLists.txt