@sporksmith 2022-09-09
Ok, I think this is ready. Sorry most of it is in one giant commit. I'd created a lot of smaller commits as I went, but ended up going back and redoing and fixing things a lot. There's a lot of interaction between different parts, s.t. using fixup commits to try to create a clean history of small commits would have been a fair bit of work. I ended up just breaking off a few small independent commits instead and putting the rest in one big one.