@sporksmith 2022-09-09
Some ad-hoc benchmarks on my machine to validate things didn't get worse:

Clean release build: 55s -> 54s
Nop incremental release build: 4s -> 2s
touch thread.c; incremental release build: 4s -> 21s

Mostly ok; incremental builds get a bit slower. I think this is because the C code is now only incremental at the crate level. When any file inside a crate is touched, we always rebuild all the C files in that crate and relink (though I think the compiled rust files within the affected crate should still be reused)