@sporksmith 2022-09-27
In addition to now always recompiling the C code, we unfortunately lose cmake's generation of compile_commands.json, which vscode and other tools use for analysis. I've verified that it can be generated by running the whole build under [bear](https://github.com/rizsotto/Bear), but haven't worked out how to get vscode to use the resulting file yet.

I poked around some more to see if we could solve the main problem without this drastic overhaul. I thought maybe using `cargo rustc`, which allows passing flags only to the "top level" without propagating to dependencies, (instead of `cargo build`) to compile the test binaries, and only passing the link flags there, might solve the problem of dependencies getting rebuilt with different build hashes, but it didn't. Unfortunately what cargo incorporates into the build hash is intentionally opaque, so I wasn't able to dig in further.

Ultimately I still think this PR is the best way forward.